+++ src/main/java/atdd/station/StationRepository.java
@@ -2,5 +2,5 @@ package atdd.station;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface StationRepository extends JpaRepository<Station, Integer> {
+public interface StationRepository extends JpaRepository<Station, Long> {
 }
